Add a base class for tests
